+++ src/sql/network/network_epoch.sql
@@ -1,0 +1,4 @@
+SELECT e.no AS "epoch"
+FROM epoch e
+  ORDER BY e.no DESC
+LIMIT 1